requirements.txt
click==8.1.3
pathlib==1.0.1
Pillow==9.2.0
tomli==2.0.1
tqdm==4.64.0
